<fix>(p2p): fix message timeout and unrecognized when p2p message forward (#3230)

+++ bcos-gateway/bcos-gateway/libp2p/P2PMessage.cpp
@@ -255,7 +255,7 @@ bool P2PMessage::tryToCompressPayload(std::shared_ptr<bytes> compressData)
     return true;
 }
 
-ssize_t P2PMessage::decodeHeader(bytesConstRef _buffer)
+int32_t P2PMessage::decodeHeader(bytesConstRef _buffer)
 {
     int32_t offset = 0;
